requirements.txt
aiodns==3.0.0
aiofile==3.7.2
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
caio==0.9.3
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptofeed==2.1.1
frozenlist==1.2.0
idna==3.3
multidict==5.2.0
order-book==0.4.1
pycares==4.1.2
pycparser==2.21
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
typing-extensions==4.0.1
urllib3==1.26.7
websockets==10.1
yapic.json==1.7.2
yarl==1.7.2
